requirements.txt
# hyperlpr==0.0.2
opencv-python==4.5.5.62
pillow==9.0.1
numpy
pyyaml